requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.34
langchain-huggingface==0.0.3
langchain-core==0.1.45
langgraph==0.0.48
huggingface-hub==0.22.2
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pypdf==4.1.0
python-dotenv==1.0.1
torch==2.2.2
transformers==4.39.3